+++ husky_components/datagrid/main.js
@@ -213,6 +213,9 @@
             },
 
             templates = {
+                errorIcon: [
+                    '<span id="error-icon" class="fa fa-exclamation-triangle"></span>'
+                ].join(''),
                 radio: [
                     '<div class="custom-radio custom-filter">',
                     '   <input name="radio-<%= columnName %>" type="radio" class="form-element" <% if (checked) { print("checked")} %>/>',
@@ -346,14 +349,30 @@
             },
 
             /**
-             * raised when save of data failed
+             * Raised when save of data failed.
+             *
              * @event husky.datagrid.data.save.failed
+             *
+             * @param {String} jqXHR status
+             * @param {String} text status
+             * @param {String} error thrown
+             */
+            DATA_SAVE_FAILED = function() {
+                return this.createEventName('data.save.failed');
+            },
+
+            /**
+             * Raised when loading of data failed.
+             *
+             * @event husky.datagrid.loading.failed
+             *
+             * @param {String} jqXHR
              * @param {String} text status
              * @param {String} error thrown
              *
              */
-            DATA_SAVE_FAILED = function() {
-                return this.createEventName('data.save.failed');
+            LOADING_FAILED = function() {
+                return this.createEventName('loading.failed');
             },
 
             /**
@@ -657,6 +676,9 @@
                 this.$element = this.sandbox.dom.$('<div class="husky-datagrid"/>');
                 this.$el.append(this.$element);
 
+                // add error icon
+                this.$element.append(templates.errorIcon);
+
                 this.dataGridWindowResize = null;
 
                 this.sort = {
@@ -764,9 +786,14 @@
                             params.success(response);
                         }
                     }.bind(this))
-                    .fail(function(status, error) {
-                        this.sandbox.logger.error(status, error);
-                    }.bind(this));
+                    .fail(this.loadingFailedHandler.bind(this));
+            },
+
+            loadingFailedHandler: function(jqXHR, textStatus, error) {
+                this.stopLoading();
+                this.showErrorIcon();
+
+                this.sandbox.emit(LOADING_FAILED.call(this), jqXHR, textStatus, error);
             },
 
             /**
@@ -936,9 +963,7 @@
                             params.success(response);
                         }
                     }.bind(this))
-                    .fail(function(status, error) {
-                        this.sandbox.logger.error(status, error);
-                    }.bind(this));
+                    .fail(this.loadingFailedHandler.bind(this));
             },
 
             /**
@@ -1053,6 +1078,20 @@
                         }
                     }
                 ]);
+            },
+
+            /**
+             * Displays the error icon
+             */
+            showErrorIcon: function() {
+                this.$find('#error-icon').show()
+            },
+
+            /**
+             * Hides the error icon
+             */
+            hideErrorIcon: function() {
+                this.$find('#error-icon').hide()
             },
 
             /**
@@ -2005,7 +2044,8 @@
                         url: this.data.links.self.href,
                         success: function() {
                             this.sandbox.emit(UPDATED.call(this));
-                        }.bind(this)
+                        }.bind(this),
+                        fail: this.loadingFailedHandler.bind(this)
                     });
                 }
             },
@@ -2031,7 +2071,8 @@
                         url: url,
                         success: function() {
                             this.sandbox.emit(UPDATED.call(this));
-                        }.bind(this)
+                        }.bind(this),
+                        fail: this.loadingFailedHandler.bind(this)
                     });
                 }
             },
@@ -2060,7 +2101,8 @@
                             url: url,
                             success: function(data) {
                                 this.sandbox.emit(UPDATED.call(this), data);
-                            }.bind(this)
+                            }.bind(this),
+                            fail: this.loadingFailedHandler.bind(this)
                         });
                     }
                 }
@@ -2114,7 +2156,8 @@
                         url: url,
                         success: function() {
                             this.sandbox.emit(UPDATED.call(this));
-                        }.bind(this)
+                        }.bind(this),
+                        fail: this.loadingFailedHandler.bind(this)
                     });
                 }
             },